Move the clean xapian query function to gn2 helper functions.
